make the NullTime type an alias of sql.NullTime
